rename web => server
