+++ main/web/server/query.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package web
+package server
 
 import (
 	"encoding/json"
